moved FunLibInfo in library analysis and deleted the other class because it was only used in the callgraph
